Add support for user directories
This commits separates the C3 data files from the user files, such as configurations, savegames, maps and custom scenario stuff.

Upon first launch with this build, the game will ask whether the user wants to specify a directory and will also, in some cases recommend one.

Base options are:
- Default (the current way, everything is thrown into the Caesar 3 directory)
- Default, with subdirectories - user files are still stored in Caesar 3, but in their own subdirectories
- Recommended - a platform-specific recommendation. Depending on the system, there may not be a recommendation)
- Custom - a user specified directory. depending on the system, this may not be available

The game will also check if the provided directory is writeable and warn the user if it isn't.

When changing the user directory, the user will be asked if he wants to copy their files to the new location and whether to overwrite files that already exist on the new location.

Because this is a big change, there may be some bugs left. I tested most of the stuff, but I may have missed something.