memory_block: Fix order of arguments to calloc()
calloc() accepts two arguments, nmemb and size, which are multiplied
together. nmemb is supposed to be the number of elements, and size the
size of each element, but we're passing the size first, then the number
of bytes.

Because we use sizeof(), this now triggers a warning in gcc:
/home/david/Development/augustus/src/core/memory_block.c: In function ‘core_memory_block_init’:
/home/david/Development/augustus/src/core/memory_block.c:7:35: warning: ‘calloc’ sizes specified with ‘sizeof’ in the earlier argument and not in the later argument [-Wcalloc-transposed-args]
    7 |     block->memory = calloc(sizeof(char), initial_size);
      |                                   ^~~~
/home/david/Development/augustus/src/core/memory_block.c:7:35: note: earlier argument should specify number of elements, later size of each element

So, while I can't see how this could actually cause an issue, let's get
rid of the warning.