Mars curse invasions are now correctly random on custom maps
`scenario_invasion_start_from_mars` did not check whether the
current map was custom or not, so it applied the same code path for all
maps even though it was originally designed for campaign only.

Now on custom maps, it should correctly use an alternate code path that
chooses a random number between 3 and 9 for the invasion force.
Currently there is no distinction between peaceful and military custom
maps.